requirements.txt
numpy>=1.24
opencv-python==4.8.1.78
matplotlib==3.7.5
patchify==0.2.3
Pillow>=9.5
segmentation-models==1.0.1
tensorflow==2.15.0
keras==2.15.0
protobuf==3.20.3
scikit-learn==1.3.2
